updated all time leaderboard service with redis caching

+++ src/main/java/com/qlish/qlish_api/dto/LeaderboardEntry.java
@@ -9,5 +9,5 @@ import lombok.Setter;
 @AllArgsConstructor
 public class LeaderboardEntry {
     private String profileName;
-    private Long points;
+    private Integer points;
 }
